requirements.txt
# EURO SECURITY - Production Requirements
# Sistema de Gestión Empresarial con GPS Tracking y PWA

# Core Django
Django==5.2.6
django-extensions==3.2.3
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-widget-tweaks==1.5.0
django-filter==23.5

# Database & Models
django-import-export==3.3.1

# Images & Files
Pillow==10.1.0

# Reports & Export
reportlab==4.0.7
openpyxl==3.1.2

# Configuration
python-decouple==3.8

# Computer Vision & AI (Facial Recognition)
opencv-python-headless==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.0

# HTTP Requests (for APIs)
requests==2.31.0

# Production Server
gunicorn==21.2.0

# Database (Production)
psycopg2-binary==2.9.7

# Static Files & Storage
whitenoise==6.6.0

# Security
django-cors-headers==4.3.1

# Monitoring & Logging
sentry-sdk==1.38.0

# Development Tools (optional)
django-debug-toolbar==4.2.0
